% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Clustering result evaluation}
\usage{
evaluation(clust, label)
}
\arguments{
\item{clust}{A list (n*1) that indicate the predicted cluster membership of each example}

\item{label}{A list (n*1) that indicate the ground-truth cluster membership of each example}
}
\value{
\item{NMI}{Normalized mutual information}

\item{RI}{Rand index}

\item{F1}{F1 score}
}
\description{
Quantitatively evaluate a clustering result
}
\details{
This function quantitatively assess the accuracy of a clustering assignment by calculating three metrics, which are normalized mutual information (NMI), Rand index, and F1 score. Note, To calculate NMI, the R package "NMI" is required to be installed first.
}
\examples{
# load BCWD dataset (569 samples, 30 dimensions, 2 clusters)
data("bcwd_data")
label = data$label
data = data[, c(1:30)]

# compute similarity matrix
S = simiMatrix(data)

# view distribution of similarity scores in S with different beta values
hist(S)

# run shrinkage clustering
set.seed(10)
clust = SuperCluster(s=S,w=100,k=20,iter=500,random=1)
clust_membership = clust$c[,1]

# use a confusion tablecompare the clustering solution with true membership
table(label,clust_membership)

# quantitatively assess the accuracy of the clustering result
eval_scores = evaluation(clust_membershipt, label)
eval_scores$NMI # Normalized mutual information
eval_scores$RI # Rand index
eval_scores$F1 #F1 score

# generate a PCA plot to visualize the clustering result
scplot(data, clust_membership)

## Note: scplot is based on ggbiplot package. Use the command below to install ggbiplot
# library(devtools)
# install_github('ggbiplot','vqv')
# library(ggbiplot)
}
\references{
Hu, Chenyue W., Hanyang Li, and Amina A. Qutub. “Shrinkage Clustering: A Fast and Size-Constrained Clustering Algorithm for Biomedical Applications.” BMC Bioinformatics 19 (2018): 19. PMC. Web. 7 May 2018.

Manning CD, Raghavan P, Schütze H, et al. Introduction to information retrieval, vol. 1. Cambridge: Cambridge university press; 2008.
}
