% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCluster.R
\name{SuperCluster}
\alias{SuperCluster}
\title{Shrinkage Clustering}
\usage{
SuperCluster(s, w = NULL, k = NULL, iter = 1000, random = 1, Path = F)
}
\arguments{
\item{s}{The similarity matrix (n*n) of a given data matrix (n*k)}

\item{w}{The minimum number of members required for each cluster}

\item{k}{The maximum number of clusters allowed}

\item{iter}{The maximum number of iterations}

\item{random}{An integer specifying the number of random initializations}

\item{Path}{A boolean value determining whether or not to  track and record the number of remaining clusters in each iteration}
}
\value{
\item{c}{An n*1 vector containg the cluster membership of each data point}

\item{a}{An n*k matrix indicating the cluster assignment status of each data point and each class}

\item{score}{An indicator of convergence}
}
\description{
Partition a data matrix using Shrinkage Clustering
}
\examples{
# load BCWD dataset (569 samples, 30 dimensions, 2 clusters)
data("bcwd_data")
label = data$label
data = data[, c(1:30)]

# compute similarity matrix
S = simiMatrix(data)

# view distribution of similarity scores in S
hist(S)

# run shrinkage clustering
set.seed(10)
clust = SuperCluster(s=S,w=100,k=20,iter=500,random=1)
clust_membership = clust$c[,1]

# quantitatively assess the accuracy of the clustering result
eval_scores = evaluation(clust_membership, label)
eval_scores$NMI # Normalized mutual information
eval_scores$RI # Rand index
eval_scores$F1 #F1 score

# generate a PCA plot to visualize the clustering result
scplot(data, clust_membership)

## Note: scplot is based on ggbiplot package. Use the command below to install ggbiplot
# library(devtools)
# install_github('ggbiplot','vqv')
# library(ggbiplot)

}
\references{
Hu, Chenyue W., Hanyang Li, and Amina A. Qutub. “Shrinkage Clustering: A Fast and Size-Constrained Clustering Algorithm for Biomedical Applications.” BMC Bioinformatics 19 (2018): 19. PMC. Web. 7 May 2018.
}
