% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperCluster.R
\name{SuperCluster}
\alias{SuperCluster}
\title{Shrinkage Clustering}
\usage{
SuperCluster(s, w = NULL, k = NULL, iter = 1000, Step = 10,
  random = 1, Path = F)
}
\arguments{
\item{s}{The similarity matrix (n*n) of a given data matrix (n*k)}

\item{w}{The minimum number of members required for each cluster}

\item{k}{The maximum number of clusters allowed}

\item{iter}{The maximum number of iterations}
}
\value{
\item{c}{n*1 vector containg cluster membership}

\item{a}{n*k matrix containing cluster membership}

\item{score}{indicator of convergence}
}
\description{
Partition a data matrix using Shrinkage Clustering
}
\examples{
\dontrun{
}

}
